Remove duplicate timestamps
Fix a bug where a timestamp file with duplicate file names could cause
the timestamp of a modified file to be reverted to an earlier date.
When timestamps are read from the file, duplicates are now removed from
the list, and the most recent timestamp for a file name is used.  This
fixes the bug, and also removes the duplicates when the timestamp file
is updated.