Regenerate the main .PJ2 file
Enhance the "Prepare for Git commit" process so that when a .PJX file is
selected for the process, the text file will be (re)generated even if
the .PJX or .PJ2 is not being tracked.  This was the earlier behavior,
but was disabled with the enhancement to generate all .PJ2 files in a
repository.